DAH-1122 First pass at refactor listing details for reusability
