+++ app/javascript/modules/listingDetails/ListingDetailsNeighborhood.tsx
@@ -1,0 +1,15 @@
+import React from "react"
+import { ListingDetailItem } from "@bloom-housing/ui-components"
+
+export const ListingDetailsNeighborhood = () => {
+  return (
+    <ListingDetailItem
+      imageAlt={""}
+      imageSrc={""}
+      title={"Neighborhood"}
+      subtitle={"Location and transportation"}
+    >
+      Map
+    </ListingDetailItem>
+  )
+}
